requirements.txt
Django>=4.2.1
djangorestframework>=3.14.0
django-guardian>=2.4.0
drf-spectacular>=0.26.2
pyYAML>=6.0
pillow>=9.5.0